Break syscalls into a new module
