+++ diy-os/src/main.rs
@@ -19,7 +19,7 @@ use bootloader_api::{
     config::{Mapping, Mappings},
     entry_point, BootInfo, BootloaderConfig,
 };
-use core::{arch::asm, panic::PanicInfo};
+use core::panic::PanicInfo;
 use diy_os::{
     allocator, hlt_loop, init,
     memory::{self, BootInfoFrameAllocator},
@@ -60,47 +60,6 @@ extern "Rust" fn main(mut boot_info: &'static mut BootInfo) -> ! {
     // print("print sys call");
 
     hlt_loop();
-}
-
-/// Parameters to functions are passed in the registers rdi, rsi, rdx, rcx, r8, r9, and further
-/// values are passed on the stack in reverse order
-///
-/// rax return
-///
-///
-#[allow(unused_variables)]
-unsafe extern "sysv64" fn sys_call<Arg1, Arg2, Arg3, Arg4, Arg5>(
-    call: u64,
-    arg1: Arg1,
-    arg2: Arg2,
-    arg3: Arg3,
-    arg4: Arg4,
-    arg5: Arg5,
-) {
-    unsafe {
-        asm!("mov rax, {0}", in(reg) call);
-    }
-
-    unsafe { x86_64::instructions::interrupts::software_interrupt::<0x80>() };
-}
-
-fn print(str: &str) {
-    let len = str.len();
-    let ptr = str.as_ptr();
-
-    unsafe { sys_call::<usize, usize, (), (), ()>(0, ptr as usize, len, (), (), ()) }
-}
-
-fn add(num: usize, other: usize) -> usize {
-    unsafe { sys_call::<usize, usize, (), (), ()>(1, num, other, (), (), ()) }
-
-    let ret: usize;
-
-    unsafe {
-        asm!("mov rax, {0}", out(reg) ret);
-    }
-
-    ret
 }
 
 /// This function is called on panic.
