+++ diy-os/src/syscalls.rs
@@ -1,0 +1,8 @@
+mod syscalls;
+mod wrapper;
+
+#[allow(unused_imports)]
+pub use wrapper::*;
+
+#[allow(unused_imports)]
+pub(crate) use syscalls::system_call_handler_wrapper;
